Drastically simplifies checkpointing
Before, we cgheckpointer per-partition. This method was not robust, and duplicated code. As a result, we created tens of thousands of tasks, and the checkpointing tasks would be starved as a result. Now, we block per-checkpoint. This means that we will be slower overall (we have to wait on the long-tail per-block), but we can put better bounds on memory and restartability each operation.