Reorganizes model training
Models are now split into "prep" and "train" submodules.
The idea is that "prep" needs to know about dask and be able to work with the checkpoints
Meanwhile, "train" should only know about pytorch and how to load / use the data.
Using the existing checkpoint system, we are also going to be able to skip duplicate work.