Improves checkpointing, removes locking on process global
We had an issue with checkpointing that our json serialization wouldn't play well with numpy arrays. To combat this we are swtich to pickle serialization. Because we have to roll-our-own serialization, we can also better check that a task has completed with a __done__ file.
This new file lists out all the file paths that are supposed to accompany a bag.

Furthermore we started detecting errors when using global locks with our dask_process_global plugin. We've removed the locks and things should still be fine. The original problem was one node initializing the same data multiple times, but it seems a lot of that is taken care of with the plugin.