Improves benchmark conv. sentence classifier
The sentence classifier no longer relies on dask.
This is because I change dask_process_global to default to the local
mock worker in the case where no dask client is supplied.
This allows the conversion script to rely only on that mock.

Additionally, the embedding util has an option to allow for a progress
bar. This will not be active by default, but for scripts like this, its
very useful.