Embedding is now determined as an average of word states
The previous K-NN results were bad, and using additional quality
measures did not improve K-NN semantic relationships. Digging in deeper,
I noticed that we're using the last attention later of BERT. This is
shown in many scenarios to be not semantically relevant. Instead, we're
going to average the word-wise hidden states.

Additionally, this commit includes some cleanup in the redis config, as
well as includes ignore lines to reject the redis database files.