Implements a key-value store for intermediate helper info
MANY tasks within pymoliere can be simplified if there is a
high-performance in-memory k-v store. However, palmetto presents
challenges when using redis. Particularly, due to transparent hugepages,
it appears that even small amounts of data can cause problems.

Even if this were not a problem it would be nice to be able to save /
load _ANY_ python object between arbitrary processes on the fly.

This commit demonstrates how the pymoliere kv_store can be used to store
the inverted index of hashed points.