Key Value store bugfix
Key value store appears to be working as intended now!
The server starts on init.
Then, the clients are prepared to pickle results.
Results must be key: list[val] (which is done for you if you use
put_many).
The connections are also behind the worker-based lock, which limits each
machine to one connection.
It appears as if this strategy works for the debug network.