Prototypes a sequential embedding predictor
Modifies train_model in order to support non-classifier models (such as
minimizing the MSE between predicted and given embeddings).

Introduces a prototype of the sequential_embedding_predictor model.
Modifies setup.py to include the plotting module that we use in train.py

The goal of this sequential sentence embedding predictor would be to
take a title, and generate an abstract. In the short-term this is done
by given x-1 embeddings, predict the x'th.