Adds basic model for abstract generation
The abstract generation model is a fine-tuning of Bert such that output
hidden states can be interpreted as word-parts. I intend to train this
model such that given one sentence, predict the follow-up. Generation
will happen through gibbs sampling.

This introduces the basic model and two tests (forward and backward).
This ensures that the abstract generator is plugged in properly to the
fine tuning layers.