Major overhaul of train_model. Introduces abstract gen main
WARNING! THIS CHANGE BREAKS COMPATIBILITY WITH THE SENTENCE CLASSIFIER

This change introduces the main training function for the abstract
generation utility. This module learns to predict the following sentence
given the current one. We do that by looking at sentence-pairs. The
second sentence of each pair is lightly modified in order to fuel the
prediction.

The train_model function is drastically overhauled. now allows for
multiple functional callbacks, and can support different types of
models.  The goal here is to have one unified set of training code that
performs per-epoch loops, and calculates plots + metrics, but doesn't
get in the way of customization.

Furthermore, the random sampling util has been moved to the file_util
module