introduces sentence generation code
given a pretrained abstract generator and an input sentence, output a
couple of sentences that are likely to go along.

This includes a test to make sure the generator produces "something".

Adds code to the abstract generator main to load a model provided one
exists.

ToDo: We need to train more than the last layer