Properly uses mask parameters for bert model
There are a number of masks that we need to set in order to use the
bert model correctly.

First, we need a token_type_id mask. This mask is 0 for all tokens in
the first sentence, and 1's for all tokens in the second.

We also need the attention mask, which is 1 for all tokens that are not
padding, and 0 otherwise.