Update 2019.10.4
Removes dependency on partd and redis in favor of mongodb.
Removes redis start scripts.
Alters the database such that we are no longer going to use graph and
non-graph keys. Instead, we're going to use simpler keys (first letter
denotes type) and then collections to separate out those types.
The graph is stored in the "graph" collection.

Furthermore the training script is altered slightly to allow for
distributed loading wherein each process takes a fraction of the input
files. Output is only produced from rank 0.