Adds validation and global metric tracking to abstract gen
The idea is that we need to have some consistent idea of the model
performance.

The validation metric is going to be based on whole-cloth generation,
similar to what we do in the test.

We also use horovod to properly coordinate an average over batches over
ranks.

Lastly, we clean up stdout so that we only see what we need to see.