Allows abstract_generator to replace tokens during training
We now introduce a probability that the abstract generator randomly
replaces tokens in the second sentence during training.
This falls under the suggestions of "Bert has a mouth..."
Furthermore, we increase the per-epoch batch size, as a way to improve
fine-tuning layers during training.
Because of this, we re-enable checkpointing, this time doing so every 10
epochs.