Abstract generator tweaks necessary for successful training
I observed that without gradient clipping, that the abstract generator
would never learn anything. Additionally, there appears to be an
off-by-one error somewhere in the tokenizer, as it seems as we are
adding one-to-many tokens during training.

While we're still going to debug this, the changes reflected in this
commit are nessesary to start w prototyping