Construction outputs json text instead of nxgraph
previously, the network construction process output nxgraphs for each of
the edges releases. This is suboptimal because we would rather just
output a format that can be directly read by the create_lookup_table
tool.

Now, we will also need to change the graph to ptbg tool to accept this
format.